requirements.txt
alembic==1.12.0
altair==5.1.2
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
branca==0.7.0
cachetools==5.3.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distlib==0.3.7
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.60
filelock==3.12.4
folium==0.14.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.0
identify==2.5.30
idna==3.4
importlib-metadata==6.8.0
inflect==7.0.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nodeenv==1.8.0
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
Pillow==10.1.0
platformdirs==3.11.0
pre-commit==3.5.0
prompt-toolkit==3.0.41
protobuf==4.24.4
ptvsd==4.3.2
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.49
st-pages==0.4.5
stack-data==0.6.3
streamlit==1.28.0
streamlit-authenticator==0.2.3
streamlit-folium==0.16.0
streamlit-js-eval==0.1.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
virtualenv==20.24.5
watchdog==3.0.0
wcwidth==0.2.12
zipp==3.17.0
zxcvbn-python==4.4.24
